safari-600.7-branch/LayoutTests/inspector-obsolete/utilities-expected.txt

This test checks Web Inspector utilities.


Running: binaryIndexOfTest

Running: lowerBoundTest

Running: upperBoundTest


Running: qselectTest
Array: []
Reference: {}
Actual:    {}
Array: [0]
Reference: {"min":0,"median":0,"max":0}
Actual:    {"min":0,"median":0,"max":0}
Array: [0,0,0,0,0,0,0,0]
Reference: {"min":0,"median":0,"max":0}
Actual:    {"min":0,"median":0,"max":0}
Array: [4,3,2,1]
Reference: {"min":1,"median":3,"max":4}
Actual:    {"min":1,"median":3,"max":4}
Array: [1,2,3,4,5]
Reference: {"min":1,"median":3,"max":5}
Actual:    {"min":1,"median":3,"max":5}
Array: [-1,3,2,7,7,7,10,12,3,4,-1,2]
Reference: {"min":-1,"median":4,"max":12}
Actual:    {"min":-1,"median":4,"max":12}

Running: sortRangeTest
